Revert "add home page view and navigation to player"
