+++ package.json
@@ -36,7 +36,6 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "bfast": "^6.0.0-alpha.18",
     "dot-prop": "^6.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
