--- public/data.json
@@ -1,20 +1,0 @@
-[
-    {
-        "name": "CopsShop",
-        "type": "off",
-        "year": 2021,
-        "id": "bafybeihza4u4mibw727nwbj47agzp5wvbmzpyqxfxwedwuuagkte3xz4ta",
-        "cover": "./cops_shop.jpg",
-        "uploaded_by": "system",
-        "category": "actions"
-    },
-    {
-        "name": "Ni Noma",
-        "type": "off",
-        "year": 2021,
-        "id": "bafybeihxnmxpahmgiesnkqkq3fl4y7atboqsxcj3555c64ceyq5v3esasq",
-        "cover": "./ni-noma-cut.jpeg",
-        "uploaded_by": "system",
-        "category": "bongo"
-    }
-]
