+++ src/pages/WatchPage.js
@@ -1,7 +1,30 @@
+import ReactHlsPlayer from 'react-hls-player';
 import React, {useEffect, useState} from "react";
 import {useQuery} from "../utils/url";
 import {useNavigate} from 'react-router-dom'
-import VideoPlayer from "../components/VideoPlayer";
+// bafybeihza4u4mibw727nwbj47agzp5wvbmzpyqxfxwedwuuagkte3xz4ta
+function HlsPlayer({v, s}) {
+    const playerRef = React.useRef();
+    useEffect(() => {
+        // playerRef.current.play()
+    }, [])
+    return (
+        <>
+            <ReactHlsPlayer
+                src={'https://infura-ipfs.io/ipfs/'+v+'/master.m3u8'}
+                autoPlay={true}
+                controls={true}
+                width="100%"
+                height="auto"
+                hlsConfig={{
+                    enableWorker: true,
+                    progressive: true,
+                    startPosition: s,
+                }}
+                playerRef={playerRef}/>
+        </>
+    )
+}
 
 function WatchPage() {
     const navigate = useNavigate();
@@ -19,10 +42,10 @@ function WatchPage() {
             return;
         }
         navigate("/", {replace: true});
-    }, [cid, navigate, time])
+    }, [cid, s])
     return (
         <div className='flex justify-center content-center vh-100'>
-            {v ? <VideoPlayer v={v} s={s}/> : null}
+            {v ? <HlsPlayer v={v} s={s}/> : null}
         </div>
     )
 }
